Separate presets, construction in `make_data_sets`
The definitions of the actual preset configurations and the code that instantiated `DataSet` objects with those configurations were previously one and the same, with a lot of repetitive boilerplate. This attempts to separate out the actual preset configurations from the code that creates DataSets, and in so doing make the module a little easier to read and to modify.

Presets are now stored in a dict named `PRESETS` and `make_data_sets()` does nothing beyond picking the appropriate presets and sending them as arguments to the `DataSet` constructor.